+++ src/main/java/com/humanup/matrix/qcm/dao/AnswerDAO.java
@@ -7,7 +7,9 @@ import org.springframework.data.repository.CrudRepository;
 import java.util.List;
 
 public interface AnswerDAO extends CrudRepository<Answer, Long> {
-    Answer findAnswerByEmailPersonAndChoiceId(String emailPerson, int choiceId);
+
+  //  @Query("SELECT a FROM Answer a WHERE lower(a.emailPerson) like %:emailPerson% and a.choice.choiceId = :choiceId ")
+    Answer findAnswerByEmailPersonAndChoiceId(String emailPerson, Long choiceId);
     List<Answer> findAll();
     Answer findById(long id);
     @Query("SELECT a FROM Answer a WHERE lower(a.emailPerson) like %:emailPerson% ")
